backend/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
pydantic[email]
